% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check-data-before-rbind-function.R
\name{check.data.before.rbind}
\alias{check.data.before.rbind}
\title{Check data before rbind}
\usage{
check.data.before.rbind(df0,df)
if(is.null(df)){df <- df0} {df <- rbind(df,df0)}
}
\arguments{
\item{df.small}{data frame to be rbinded with a larger dataframe (df)
'@param df the larger data frame}

\item{action}{what to do in case of error (="Stop.if.error" is default)}
}
\value{
nothing (printed statement)
}
\description{
Help identify problems before blindly trying to rbind two dataframes.
}
\author{
Claus E. Andersen
}
