% Generated by roxygen2 (4.0.1): do not edit by hand
\name{clan.install}
\alias{clan.install}
\title{Function to ease the installation of important R packages}
\usage{
clan.install()
}
\value{
Side effect is that selected iportant libraries are installed
}
\description{
Function to ease the installation of important R packages
}
\author{
Claus E. Andersen
}

