% Generated by roxygen2 (4.0.1): do not edit by hand
\name{coefficients.ca}
\alias{coefficients.ca}
\title{Extract fitted paarmeters and uncertainties from a fitted lm-model.}
\usage{
\method{coefficients}{ca}(fm)
}
\arguments{
\item{fm}{a fitted model from lm}
}
\value{
dataframe with coefficients and uncertainties
}
\description{
Extract fitted paarmeters and uncertainties from a fitted lm-model.
}
\author{
Claus E. Andersen
}

