% Generated by roxygen2 (4.0.1): do not edit by hand
\name{create.date.string}
\alias{create.date.string}
\title{Create date string}
\usage{
create.date.string(x=c("05.05.2014","06.05.2014","07.05.2014"),format="6")
}
\arguments{
\item{format:}{"1","2" .. "7" and "99"}
}
\value{
string or list
}
\description{
Create data string of different formats. The
function is vectorized.
}
\author{
Claus E. Andersen
}

