% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clanToolsFunctions.R
\name{dayno.calc}
\alias{dayno.calc}
\title{Calculate the difference between two time stamps}
\usage{
dayno.clock(c("12:00","23:59","00:00:59","12.00.30","23x30"))
}
\arguments{
\item{Tday}{is the date (e.g. "31.01.2001")}

\item{Ttime}{is the clock (e.g. "00:00")}

\item{ref.date}{is the reference date (e.g. "31.12.2000")}

\item{ref.time}{is the reference time (e.g. "00:00")}
}
\value{
The difference between the given time stamp and the reference time stamp in number of days day numbers
}
\description{
Dates and wall-clock times are used as input. The output is in number of days .
Note: 1) Any delimiter will do, 2) seconds are option
}
\author{
Claus E. Andersen
}
