% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clanToolsFunctions.R
\name{dayno.clock.reversed}
\alias{dayno.clock.reversed}
\title{Convert from  fraction of day to wall-clock time.}
\usage{
dayno.clock.reversed(c(0.0,0.3,4444.5))
}
\arguments{
\item{Ttime}{is the time to be converted}
}
\value{
vector with fractional day numbers
}
\description{
Convert from fraction of day to wall-clock time.
Full days are ignored (so 0.5, 1.5, and 5573476.5 all convert to "12:00").
Likewise, for example, 0.0 ane 1.0 both convert to "00:00" (Not "24:00").
Sample call: dayno.clock.reversed(c(0.0,0.3,4444.5)) (returns: "00:00",""07:12","12:00")
}
\author{
Claus E. Andersen
}

