% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clanToolsFunctions.R, R/file-functions.R
\name{dayno.from.fileinfo}
\alias{dayno.from.fileinfo}
\title{Compute dayno from fileinfo output}
\usage{
See also: \link{dayno.calc}

none
}
\arguments{
\item{ftime}{is the output from file.info("some-file-name").}

\item{format}{is reserved for future use (currently there is only one format).}
}
\value{
vector with dayno

dayno
}
\description{
Calculate dayno for date output from fileinfo.
The function assumes the following format of the fileinfo

# The function assumes the following format of the fileinfo
times: "2020-07-15 16:38:28 CEST"
Sample call: 
  xx <- file.info(c("alanine-mass-001.txt"))
  xx.dayno <- dayno.from.fileinfo(xx$mtime)
}
\author{
Claus E. Andersen

Claus E. Andersen
}
