% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/English-months-function.R
\name{English.months}
\alias{English.months}
\title{Print names of months in English}
\usage{
English.months(1)
}
\arguments{
\item{i}{vector with numbers between 1 and 12.}
}
\value{
string vector
}
\description{
Convert numbers from 1 to 12 into "January", "February" etc.
This version handles vectorized input, and NA's and NULL.
See also: \link{dayno.calc} \link{conv.ref.date} \link{dayno.clock.reversed} \link{dayno.clock} \link{get.date.range}
}
\author{
Claus E. Andersen
}
