% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-functions.R
\name{extract.drive}
\alias{extract.drive}
\title{Extract drive letter from full path (not vectorized)}
\usage{
none
}
\arguments{
\item{fn}{is the full path to a file such as "c:/data/catest.txt".}

\item{drive.delimeter}{is normally ":".}
}
\value{
list(org=fn, drive=before, rest=after)
}
\description{
fn the   full path to a single file. This function identifies the drive letter (if there is one).
There is a vectorized version of this function (see link below).
Three situations
(1) None :  No drive is given.  Drive=""          Rest=x
(2) One  :  The drive is given. Drive=before      Rest=after
(3) Two+ :  Nonsense syntax.    Drive=before.last Rest=after.last
See also \link{extract.drive.vec} and \link{split.filename}.
}
\author{
Claus E. Andersen
}
