% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-functions.R
\name{extract.drive.vec}
\alias{extract.drive.vec}
\title{Extract drive letter from full path (vectorized)}
\usage{
none
}
\arguments{
\item{fn.vec}{is a vector of full paths to files such as "c:/data/catest.txt".}
}
\value{
list(org=fn, drive=before, rest=after)
}
\description{
fn is a vector full paths to files. This function identifies the drive letter (if there is one).
Three situations
(1) None :  No drive is given.  Drive=""          Rest=x
(2) One  :  The drive is given. Drive=before      Rest=after
(3) Two+ :  Nonsense syntax.    Drive=before.last Rest=after.last
See also \link{extract.drive.vec} and \link{split.filename}.
}
\author{
Claus E. Andersen
}
