% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clanToolsFunctions.R
\name{extract.first.number}
\alias{extract.first.number}
\title{Extract the first number from a string}
\usage{
extract.first.number(c("aa+45", "hello9abe999"))
}
\arguments{
\item{x}{is a vector of string values}
}
\value{
A list with the numbers ($number) and the remaining strings after the number ($rest.string).
}
\description{
Extract the first numbers from each individual string element in a vector.
}
\author{
Claus E. Andersen
}

