% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clanToolsFunctions.R
\name{extract.given.number}
\alias{extract.given.number}
\title{Extract the N'th number from a string}
\usage{
extract.given.number(c("aa1","bbb tet 11 22 33"),3)
}
\arguments{
\item{x}{is a vector of string values}

\item{N}{= 1 means the first number, N = 2 means the second number etc.}
}
\value{
A vector with the numbers.
}
\description{
Extract the N'th numbers from each individual string element in a vector.
}
\author{
Claus E. Andersen
}
