% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-dots-functions.R
\name{extract.last.part.of.string}
\alias{extract.last.part.of.string}
\title{Extract the tail part of a string}
\usage{
extract.last.part.of.string(c("Hello","khdskhdsjkdkahdshkdhkahskaksdjh"),n=5)
}
\arguments{
\item{x}{string to be modified}

\item{n}{desired length of string}
}
\value{
string with given length taken from the tail
}
\description{
This function extracts the last n characters
from a string.
}
\author{
Claus E. Andersen
}
