% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/English-months-function.R
\name{get.date.range}
\alias{get.date.range}
\title{Make string with range of dates (eg. for calibration certificates)}
\usage{
#' get.date.range(x=c("01-01-2020","31-01-2020"))
}
\arguments{
\item{x}{vector with dates in the format "31-01-2020".}
}
\value{
string with human readable text.
}
\description{
Compute the range of dates and present results using English months, like
"January 01, 2020 -- January 31, 2020". If there is only one date, then
the output will be something like "January 01, 2020 (one day only)".
 
Example:

get.date.range(x="31-01-2020")

get.date.range(x=c("01-01-2020","01-01-2020"))

get.date.range(x=c("01-01-2020","31-01-2020"))
}
\author{
Claus E. Andersen
}
