% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row-wise-computations-functions.R
\name{get.selected.data}
\alias{get.selected.data}
\title{Improved function for computation of row means (row-wise means).}
\usage{
df \%>\% mutate(V.base = get.selected.data(.,V.base.use.vec))
}
\arguments{
\item{df}{data frame}

\item{name.vec}{names of rows.}

\item{null.default}{(e.g. 0.0).}
}
\value{
string vector
}
\description{
Useful function for extracting data and combining data based on a vector of column names in a dataframe.
If name.vec is NULL, then assign the value given in null.default
Used in: step42-electrometer-cal-single-file-009-6

Sample use: read.table(fn.full,sep=";",header=TRUE,quote = "\"",) \%>\%

           filter(step.i \%in\% use.these.steps) \%>\%
           
           mutate(V.base = get.selected.data(.,V.base.use.vec)) \%>\%
           
           mutate(V.raw = get.selected.data(.,V.use.vec)) \%>\%
           
           mutate(V.ref = V.raw - V.base)
}
\author{
Claus E. Andersen
}
