% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row-wise-computations-functions.R
\name{get.selected.data.sd}
\alias{get.selected.data.sd}
\title{Extract columns from dataframe and compute row standard deviations (row-wise standard deviations).}
\usage{
df \%>\% mutate(V.raw.sd = get.selected.data.sd(.,V.use.vec))
}
\arguments{
\item{df}{data frame}

\item{name.vec}{names of rows.}

\item{null.default}{(e.g. 0.0).}
}
\value{
string vector
}
\description{
Useful function for extracting data and combining data based on a vector of column names in a dataframe.
Useful function for extracting data and combining data based on a vector of column names in a dataframe.
See \link{get.selected.data} and \link{rowMeans.ca}.
Here we compute (row by row) the standard deviation of the selected colums
}
\author{
Claus E. Andersen
}
