% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-summary-functions.R
\name{last.element.character}
\alias{last.element.character}
\title{Find last element and return as character.}
\usage{
last.element.character(1:10)
}
\arguments{
\item{x}{input vector}
}
\value{
string (single value)
}
\description{
Find last element and return as character.
See also \link{first.element.character}, \link{mean.for.numerics}, 
\link{min.for.numerics}, \link{max.for.numerics}, and \link{number.of.nas}.
}
\author{
Claus E. Andersen
}
