% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clanToolsFunctions.R
\name{leading.blanks}
\alias{leading.blanks}
\title{Add leading blanks.}
\usage{
leading.blanks(xx.LATEX,"x") See example in body of function.
}
\arguments{
\item{df}{is dataframe}

\item{blank.char}{is the character to be added}

\item{extra.char}{is an extra character}

\item{min.width}{is the minimum width}

\item{max.width}{is the maximum width}

\item{ignore.first}{is gives special treatment to first column}

\item{trace}{equal to TRUE produces output during execution}
}
\value{
A vector with the rounded numbers.
}
\description{
Add leading blanks. This is to align data in Latex tables. 
Changed from: cc.width <- nchar(as.character(cc))
to: cc.width <- nchar(as.character(cc),type="width")
Why? To avoid problems with NA's in dataframes in txtplot.
Main purpose: to align data in Latex tables.
The ignore.first parameters controls the first row (often a row with names).
For example:
  xx <- data.frame(sample(c(1,1055500,10),10,replace=T),sample(c("a","aba","sfgsdfgfhgd"),10,replace=T))
  xx.LATEX <- create.latex.table(xx)
  yy <- leading.blanks(xx.LATEX,"x")
  yy
  write.table(yy,"xx.txt",sep=" ",dimnames.write = F)
}
\author{
Claus E. Andersen
}
