% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clanToolsFunctions.R
\name{leading.zeros}
\alias{leading.zeros}
\title{Add zeros in string representing a number}
\usage{
leading.zeros(c(9,111),3)
}
\arguments{
\item{x}{is a vector of numeric values}

\item{digits}{is the number of digits in the final results.}
}
\value{
The vecor of characters wher zeors have been added as necessary
}
\description{
Add zeros when numbers occur in filenames and such
009 comes before 111 (cf. 9 vs. 111)
}
\author{
Claus E. Andersen
}

