% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clanToolsFunctions.R
\name{leading.zeros.to.fit}
\alias{leading.zeros.to.fit}
\title{Add zeros in string representing a number}
\usage{
leading.zeros.to.fit(c(9,111))
}
\arguments{
\item{x}{is a vector of numeric values}
}
\value{
The vecor of characters wher zeors have been added as necessary
}
\description{
Add zeros when numbers occur in filenames and such
009 comes before 111 (cf. 9 vs. 111). The number of leading zeros
is camculated automatically.
}
\author{
Claus E. Andersen
}
