% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm-extract-function.R
\name{lm.extract}
\alias{lm.extract}
\title{Extract data from fitted lm-model}
\usage{
df <- data.frame(x=1:10, y=rnorm(10))
  fm <- lm(y~x,data=df) 
  cc <- lm.extract(fm)
}
\arguments{
\item{trace}{controls output during execution}
}
\value{
list of data
}
\description{
Extract data from fitted lm-model
}
\author{
Claus E. Andersen
}
