% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-summary-functions.R
\name{max.for.numerics}
\alias{max.for.numerics}
\title{Find maximum element, if numeric}
\usage{
max.for.numerics(1:10)
}
\arguments{
\item{x}{input vector}

\item{digits}{number of digits in result}
}
\value{
string with given number of digits (single value)
}
\description{
Find maximum element, if numeric.
See also \link{first.element.character}, \link{last.element.character}, \link{mean.for.numerics}, 
\link{min.for.numerics}, and \link{mumber.of.nas}.
Use case: see CBASE-cobalt-cal-single-100.R
}
\author{
Claus E. Andersen
}
