% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-summary-functions.R
\name{mean.for.numerics}
\alias{mean.for.numerics}
\title{Find mean element, if numeric}
\usage{
mean.for.numerics(1:10)
}
\arguments{
\item{x}{input vector}

\item{digits}{number of digits in result}
}
\value{
string with given number of digits (single value)
}
\description{
Find mean element, if numeric.
See also \link{first.element.character}, 
\link{min.for.numerics}, \link{max.for.numerics}, and \link{mumber.of.nas}.
Use case: see CBASE-cobalt-cal-single-100.R
}
\author{
Claus E. Andersen
}
