% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-summary-functions.R
\name{min.for.numerics}
\alias{min.for.numerics}
\title{Find minimum element, if numeric}
\usage{
min.for.numerics(1:10)
}
\arguments{
\item{x}{input vector}

\item{digits}{number of digits in result}
}
\value{
numeric with given number of significant digits (single value)
}
\description{
Find minimum element, if numeric.
See also \link{first.element.character}, \link{last.element.character}, \link{mean.for.numerics}, 
\link{min.for.numerics}, \link{max.for.numerics}, and \link{number.of.nas}.
Use case: see CBASE-cobalt-cal-single-100.R
}
\author{
Claus E. Andersen
}
