% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clanToolsFunctions.R
\name{most.common.element}
\alias{most.common.element}
\title{Find the most common element(s) in a single vector.}
\usage{
xx <- c(NA,1:100,NA)
most.common.element(xx,1)
}
\arguments{
\item{x}{is the vector}

\item{max.number.of.elements}{is the limit elements in the return vector}
}
\value{
The most common element in the vector
}
\description{
Find the most common element(s) in a single vector
This functions also handles NULL vectors and NA-values.
}
\author{
Claus E. Andersen
}
