% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc-summary-functions.R
\name{number.of.nas}
\alias{number.of.nas}
\title{Find number of NAs in vector}
\usage{
number.of.nas(1:10)
}
\arguments{
\item{x}{input vector}
}
\value{
number (single value)
}
\description{
Find number of NA-values in vector.
See also \link{first.element.character}, \link{last.element.character}, \link{mean.for.numerics}, 
\link{min.for.numerics}, and \link{max.for.numerics}.
Use case: see CBASE-cobalt-cal-single-100.R
}
\author{
Claus E. Andersen
}
