% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clanToolsFunctions.R
\name{round.ca}
\alias{round.ca}
\title{Round off to given number of decimals}
\usage{
round.ca(rnorm(100),1)
}
\arguments{
\item{x}{is a vector of numeric  values}

\item{decimals}{is the number of requested decimals}
}
\value{
A vector with the rounded numbers.
}
\description{
Round off to given number of decimals.
}
\author{
Claus E. Andersen
}

