% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clanToolsFunctions.R
\name{round.res}
\alias{round.res}
\title{Reformat number to given resolution.}
\usage{
round.res(4.5555)
}
\arguments{
\item{x}{is a vector of numbers}

\item{resolution}{is the number of requested digits}
}
\value{
A vector of string with the rounded numbers.
}
\description{
Reformat number to given resolution.
}
\author{
Claus E. Andersen
}
