% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clanToolsFunctions.R
\name{session.info.ca}
\alias{session.info.ca}
\title{Session information}
\usage{
session.info.ca()
}
\value{
List with information. Printing data as a side effect.
}
\description{
List main information about operating system, R, and packages
}
\author{
Claus E. Andersen
}
