% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clanToolsFunctions.R
\name{signif.res}
\alias{signif.res}
\title{Use significance to reformat number to given resolution.}
\usage{
signif.res(c(1.26763,0.00123,1,0.0011,1e-19,1.1e-4,NA),2)
}
\arguments{
\item{x}{is a vector of numbers}

\item{resolution}{is the number of significant digits}

\item{resolution.zero}{is the number of fixed-number-of-digits near zero.}

\item{take.as.zero}{If a number is below the take.as.zero, then we apply the fixed-number-of-digits approach.}

\item{print.for.na}{is printed for NA-values}

\item{method}{is the method used to determine significance (default method = "signif").}
}
\value{
A vector of string with the rounded numbers.
}
\description{
Reformat number to given resolution.
}
\author{
Claus E. Andersen
}
