% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clanToolsFunctions.R
\name{substitute.char}
\alias{substitute.char}
\title{Replace a character in a string with a new character}
\usage{
substitute.char("abxdefg","x","c")
}
\arguments{
\item{st}{is the string}

\item{find.char}{is the character to be replaced}

\item{new.char}{is the new character}
}
\value{
The string with the replaced character
}
\description{
Replace a single character.
}
\author{
Claus E. Andersen
}

