% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-dots-functions.R
\name{substring.with.dots.adv}
\alias{substring.with.dots.adv}
\title{Produce a string with given length including head, tail and midsection}
\usage{
substring.with.dots.adv(c("Hello","khdskhdsjkdkahdshkdhkahskaksdjh"),n=15,reverse=!TRUE,n.last=5)
print(c("Hello ........." ,"khdskhd...aksdj"))
}
\arguments{
\item{x}{string to be modified}

\item{n}{desired length of string after manipulations}

\item{reverse}{controls if the dots are added before of after the main string}

\item{n.last}{number of characters from the tail}

\item{mid.string}{what to put between the head and the tail to indicate that some part of the mid section has been cut out.}
}
\value{
string with given length including head, tail and a midsection.
}
\description{
Sometimes we need to summarize long strings, and we need 
a bit og the head and a bit of the tail, and some indication that
we are not seeing the full string. This function does that plus it can 
pad dots or other symbons before of after the string, if it is not
long.
See also \link{table.dots} and \link{substring.with.dots}.
}
\author{
Claus E. Andersen
}
