% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-dots-functions.R
\name{substring.with.dots}
\alias{substring.with.dots}
\title{To pad dots (or other text) to tables for improved readability.}
\usage{
substring.with.dots("Hello")
}
\arguments{
\item{x}{is a vector of input values}

\item{n}{combined length (incl. dots)}

\item{reverse}{(FALSE = x then dots, TRUE = dots then x)}

\item{st}{is the dot-character (i.e. normally ".").}
}
\value{
string (vectorized)
}
\description{
Objective: truncate string to n chars and pad with dots in front
behind depending on reverse.
Example:

substring.with.dots("Hello")

gives the following output:

"Hello .............."

See also \link{table.dots}.
}
\author{
Claus E. Andersen
}
