% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-dots-functions.R
\name{table.dots}
\alias{table.dots}
\title{To pad dots (or other text) to tables for improved readability.}
\usage{
table.dots("k.elec",N=40,before=!FALSE,st=".",st.space=" ")
}
\arguments{
\item{txt}{A single string (not a vector).}

\item{N}{number of dots (e.g. 40).}

\item{before}{should be set to TRUE if the dots should preceed the txt.}

\item{st}{is the dot-string (e.g. ".").}

\item{st.space}{(e.g. " = ").}
}
\value{
string vector
}
\description{
Objective: To pad dots (or other text) to tables for improved readability. 
See also: \link{table.dots.vec} and \link{substring.with.dots}.
Sample call: 

table.dots("k.elec",N=40,before=!FALSE,st=".",st.space=" ")

table.dots("k.elec",N=40,before=FALSE,st="Error ",st.space=" = ")

Produces this output:

[1] "................................. k.elec"

[1] "k.elec = Error Error Error Error Error E"
}
\author{
Claus E. Andersen
}
