% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table-dots-functions.R
\name{table.dots.vec}
\alias{table.dots.vec}
\title{To pad dots (or other text) to tables for improved readability. This is a vectorized version.}
\usage{
table.dots("k.elec",N=40,before=!FALSE,st=".",st.space=" ")
}
\arguments{
\item{N}{number of dots (e.g. 40).}

\item{before}{should be set to TRUE if the dots should preceed the txt.}

\item{st}{is the dot-string (e.g. ".").}

\item{st.space}{(e.g. " = ").}

\item{txt}{vector of strings.}
}
\value{
string vector
}
\description{
Objective: To pad dots (or other text) to tables for improved readability.
This is a vectorized version. See also \link{table.dots}.
Sample call: 

table.dots.vec(c("k.elec","k.elec.pos","k.elec.mean"),N=40,before=FALSE,st=".",st.space=" ") 

Produces this output:

[1] "k.elec ................................."

[2] "k.elec.pos ............................."

[3] "k.elec.mean ............................"
}
\author{
Claus E. Andersen
}
