% Generated by roxygen2 (4.0.1): do not edit by hand
\name{thermistor.ohm}
\alias{thermistor.ohm}
\title{Convert from temperature to resistance using three parameters}
\usage{
see body of function.
}
\arguments{
\item{b}{a vector of three elements.}
}
\value{
the resistance in ohm
}
\description{
Conversion for NTC (e.g. 2 kohm) thermistors (Risoe parameterization)
}
\author{
Claus E. Andersen
}

