% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clanToolsFunctions.R
\name{workflow.ca}
\alias{workflow.ca}
\title{Template for graphical output}
\usage{
Do not run - just view the body of the function and use it as a template.
}
\value{
Side effect is that selected iportant libraries are installed
}
\description{
Template for graphical output using pdf, ps or png format.
# Recommended workflow for making PowerPoint reports:
(1) Print all files to consequtive png-files
(2) Import to PowerPoint using the photoalbum feature.
    This puts each plot in separate slides in the presentation.
}
\author{
Claus E. Andersen
}

