% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clanToolsFunctions.R
\name{wrline}
\alias{wrline}
\title{Write a double line}
\usage{
wrline("End")
}
\arguments{
\item{txt}{is an optional txt message that will occur on the line.}

\item{N.blanks}{is the number of blanks lines before the first dashed line}

\item{N.blanks.after}{is the number of blanks before the second dashed line}

\item{N}{is the total number of characters in the line.}
}
\value{
NULL (the side effect is that a line is printed)
}
\description{
This function can be used to separate report output
}
\author{
Claus E. Andersen
}
