% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clanToolsFunctions.R
\name{clan.install}
\alias{clan.install}
\title{Function to ease the installation of important R packages}
\usage{
clan.install()
}
\value{
Side effect is that selected iportant libraries are installed
}
\description{
Function to ease the installation of important R packages
}
\author{
Claus E. Andersen
}
