% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clanToolsFunctions.R
\name{clanTools}
\alias{clanTools}
\title{Version function for the clanTools library}
\usage{
clanTools()
}
\value{
list with information about the version and the functions in the library
}
\description{
Version function for the clanTools library
}
\author{
Claus E. Andersen
}
