% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clanToolsFunctions.R
\name{coefficients.ca}
\alias{coefficients.ca}
\title{Extract coefficients from a fitted linear model (lm)}
\usage{
fm <- lm(x~y, data=df)
cc <- coefficents.ca(fm)
print( cc$coeff )
print( cc$u.coeff )
}
\arguments{
\item{fm}{is a fitted model from lm}
}
\value{
A dataframe with coefficients and uncertainties. coeff is the
fitted values and u.coeff is the accociated uncertainties.
}
\description{
This function helps get easy access to the uncertainties associated with the fit.
}
\author{
Claus E. Andersen
}
