% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/English-months-function.R
\name{conv.ref.date}
\alias{conv.ref.date}
\title{Compute dayno from a written date like March 22, 2018}
\usage{
#' get.date.range(x=c("01-01-2020","31-01-2020"))
}
\arguments{
\item{ref.date.vec}{vector with dates in the format "March 22, 2018".}

\item{ref.time.vec}{vector (or just a single value) with times in the format "12:00".}

\item{zero.date}{in the format "01.01.2014".}

\item{zero.time}{in the format "12:00".}
}
\value{
dateframe with detailed info (see example)
}
\description{
Convert date strings of the type March 22, 2018 to day numbers
since 01.01.2014 or some other zero date (reference). 
Sample call: 

 conv.ref.date("March 22, 2018, 12:00")
  
 conv.ref.date(c("March 22, 2018", "March 21, 2018", "March 22, 2019"),c("12:00"),
  zero.time="00:00",zero.date="22.03.2018") 
 
 with this output:
 
        dstr.org  d m    y       dstr dayno
 
        March 22, 2018 22 3 2018 22.03.2018   0.5
        
        March 21, 2018 21 3 2018 21.03.2018  -0.5
        
        March 22, 2019 22 3 2019 22.03.2019 365.5
}
\author{
Claus E. Andersen
}
