% Generated by roxygen2 (4.0.1): do not edit by hand
\name{create.latex.table}
\alias{create.latex.table}
\title{Convert a dateframe into a Latex table}
\usage{
see example in body of function.
}
\arguments{
\item{df}{is the dataframe to be converted.}

\item{pretty.func}{is a function needed to get the desired resultion for each individual column.}

\item{NL}{is the characted indicatinf new line.}

\item{col.sign}{is the character for column seperation.}
}
\value{
NULL (the side effect is that a line is printed)
}
\description{
Generate a table for Latex (with &'s and \\-line endings).
}
\author{
Claus E. Andersen
}

