% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clanToolsFunctions.R
\name{dayno.clock}
\alias{dayno.clock}
\title{Convert from wall-clock time to fraction of day}
\usage{
dayno.clock(c("12:00","23:59","00:00:59","12.00.30","23x30"))
}
\arguments{
\item{Ttime}{is the time to be converted}
}
\value{
vector with fractional day numbers
}
\description{
Convert from wall-clock time to fraction of day.
Note: 1) Any delimiter will do, 2) seconds are option
}
\author{
Claus E. Andersen
}
