% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clanToolsFunctions.R
\name{dayno.from.fileinfo}
\alias{dayno.from.fileinfo}
\title{Compute dayno from fileinfo output}
\usage{
See also: \link{dayno.calc}
}
\arguments{
\item{format}{select the format of the date (not implemented)}
}
\value{
vector with dayno
}
\description{
Calculate dayno for date output from fileinfo.
The function assumes the following format of the fileinfo
}
\author{
Claus E. Andersen
}
