% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/English-months-function.R
\name{English.months}
\alias{English.months}
\title{Print names of months in English}
\usage{
English.months(1)
}
\arguments{
\item{i}{number between 1 and 12.}
}
\value{
string
}
\description{
Convert numbers 1 to 12 into "January", "February" etc.
}
\author{
Claus E. Andersen
}
