% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clanToolsFunctions.R
\name{last.element}
\alias{last.element}
\title{Find the last element in a vector}
\usage{
xx <- c(NA,1:100,NA)
last.element(xx)
}
\arguments{
\item{x}{is the vector}

\item{na.ignore}{controls if NA-values should be ignored}

\item{N}{(default = 1) is the number of elements that should be returned}
}
\value{
The last element
}
\description{
This functions also handles NULL vectors and NA-values.
}
\author{
Claus E. Andersen
}
