% Generated by roxygen2 (4.0.1): do not edit by hand
\name{leading.blanks}
\alias{leading.blanks}
\title{Add leading blanks.}
\usage{
leading.blanks(xx.LATEX,"x") See example in body of function.
}
\arguments{
\item{df}{is dataframe}

\item{blank.char}{is the character to be added}

\item{extra.char}{is an extra character}

\item{min.width}{is the minimum width}

\item{max.width}{is the maximum width}

\item{ignore.first}{is gives special treatment to first column}

\item{trace}{equal to TRUE produces output during execution}
}
\value{
A vector with the rounded numbers.
}
\description{
Add leading blanks. This is to align data in Latex tables.
}
\author{
Claus E. Andersen
}

