% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace-token-function.R
\name{replace.token}
\alias{replace.token}
\title{To replace tokens in documents (e.g. calibration certificate templates) with a given string.}
\usage{
txt <- replace.token(txt,"TOKEN.CERTDATE","December 24, 1999")
}
\arguments{
\item{TOKEN.ID}{(e.g. "TOKEN.XYZ").}

\item{Replace.auto}{(e.g. "").}

\item{Replace.manual}{as an alternative (set to NULL).}
}
\value{
text (typically an entire document with many lines).
}
\description{
To replace tokens in in documents (e.g. calibration certificate templates) with valid input - typically generated by the software. 
The modified text is for Latex typesetting.
Sample call:

txt <- "jsjdksfkd  TOKEN.CERTDATE jsfhjhfjkhjfadsk safjsfkf"

txt <- replace.token(txt,"TOKEN.CERTDATE","December 24, 1999")
}
\author{
Claus E. Andersen
}
