% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprintf-ca-function.R
\name{sprintf.ca}
\alias{sprintf.ca}
\title{Print number to fit certain format. An alternative is used if this requires too many characters.}
\usage{
\verb{sprintf.ca( "12:00", round(6.6e42,0),max.chars=9999)}
}
\arguments{
\item{format}{is dayno.clock(c("12:00","23:59","00:00:59","12.00.30","23x30"))}

\item{number}{to print.}

\item{format.alt}{is the alternative format.}

\item{max.char}{is the maximum of characters the the format needs to fit (otherwise the alternative format is used).}
}
\value{
formated number.
}
\description{
This is a variation of sprintf, that checks if the formated results is too long.
}
\author{
Claus E. Andersen
}
