% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprintf-ca-function.R
\name{sprintf.ca}
\alias{sprintf.ca}
\title{Print number to fit certain format. An alternative is used if this requires too many characters.}
\usage{
sprintf.ca("\%.5f", round(5.1,0),max.chars=9999)
}
\arguments{
\item{format}{is the main format (e.g. "\%.5f").}

\item{number}{to print (e.g. 10).}

\item{format.alt}{is the alternative format (e.g. "\%.4e").}

\item{max.char}{is the maximum of characters the the format needs to fit (otherwise the alternative format is used).}
}
\value{
formated number.
}
\description{
This is a variation of sprintf, that checks if the formated results is too long. 
This can happen, for example, with electrometer output, where normal results are 
0.004 pC and then an overflow gives a charge of 6e47 C, which produces an output 
that ruins txtplot table output.
}
\author{
Claus E. Andersen
}
