% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clanToolsFunctions.R
\name{thermistor.degC}
\alias{thermistor.degC}
\title{Convert from NTC to temperature using three parameters}
\usage{
see body of function.
}
\arguments{
\item{b}{a vector of three elements.}
}
\value{
the temperature in degC
}
\description{
Conversion for NTC (e.g. 2 kohm) thermistors (Risoe parameterization)
}
\author{
Claus E. Andersen
}

