% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clanToolsFunctions.R
\name{thermistor.ohm}
\alias{thermistor.ohm}
\title{Convert from temperature to resistance using three parameters}
\usage{
see body of function.
}
\arguments{
\item{b}{a vector of three elements.}
}
\value{
the resistance in ohm
}
\description{
Conversion for NTC (e.g. 2 kohm) thermistors (Risoe parameterization)
}
\author{
Claus E. Andersen
}
