% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim-whitespace-function.R
\name{trim.whitespace}
\alias{trim.whitespace}
\title{Remove write space at the beginning and end of a string}
\usage{
x <- trim.whitespace("    uyiyy uiyiyiuy y        ")
}
\value{
trimmed string
}
\description{
Trim a string for white space.
}
\author{
Claus E. Andersen
}
